refactor: merge verify email to user_handler
