--- internal/handler/user_handler_verify_email.go
@@ -1,46 +1,0 @@
-package handler
-
-import (
-	"github.com/PitiNarak/condormhub-backend/internal/dto"
-	"github.com/PitiNarak/condormhub-backend/pkg/apperror"
-	"github.com/go-playground/validator"
-	"github.com/gofiber/fiber/v2"
-)
-
-// VerifyEmail godoc
-// @Summary Verify email
-// @Description Verify email
-// @Tags user
-// @Accept json
-// @Produce json
-// @Param user body dto.VerifyRequestBody true "token"
-// @Success 200 {object} dto.SuccessResponse[dto.TokenWithUserInformationResponseBody] "email is verified successfully"
-// @Failure 400 {object} dto.ErrorResponse "your request is invalid
-// @Failure 401 {object} dto.ErrorResponse "your request is unauthorized"
-// @Failure 500 {object} dto.ErrorResponse "system cannot verify your email"
-// @Router /user/verify [post]
-func (h *UserHandler) VerifyEmail(c *fiber.Ctx) error {
-	body := new(dto.VerifyRequestBody)
-
-	if err := c.BodyParser(body); err != nil {
-		return apperror.BadRequestError(err, "your request is invalid")
-	}
-	validate := validator.New()
-
-	if err := validate.Struct(body); err != nil {
-		return apperror.BadRequestError(err, "your request body is incorrect")
-	}
-	accessToken, user, err := h.userService.VerifyUser(c.Context(), body.Token)
-	if err != nil {
-		return err
-	}
-
-	data := dto.TokenWithUserInformationResponseBody{
-		AccessToken:     accessToken,
-		UserInformation: *user,
-	}
-
-	res := dto.Success(data)
-
-	return c.Status(fiber.StatusOK).JSON(res)
-}
